+++ source/runtime/blocks/control.cpp
@@ -100,7 +100,7 @@ SCRATCH_BLOCK(control, create_clone_of) {
     const int sourceIndex = (Scratch::sprites.size() - 1) - targetSprite->layer;
     auto it = Scratch::sprites.insert(Scratch::sprites.begin() + sourceIndex + 1, clonedSprite);
 
-    for (int i = sourceIndex + 1; i < Scratch::sprites.size(); i++) {
+    for (size_t i = sourceIndex + 1; i < Scratch::sprites.size(); i++) {
         Scratch::sprites[i]->layer = (Scratch::sprites.size() - 1) - i;
     }
